requirements.txt
ShopifyAPI==10.0.0
python-dotenv==0.21.0
